fix: resolve all clippy warnings across workspace
- Collapse nested if-let patterns (mir/optimize, lint/unused_variable, lsp/symbols)
- Merge identical branches in parser visibility handling
- Use strip_prefix instead of manual slicing (dom)
- Add Default impl for PowerMetrics
- Remove unnecessary parentheses (wasm func_section)
- Remove unused lifetime parameter (llvm infer_value_llvm_type)
- Use enumerate instead of manual indexing (uuid)
- Add type alias for complex callback type (tea ResponseCallback)
- Allow clippy lints where intentional: should_implement_trait on
  abstract domain ops (absinterp), too_many_arguments on lowering
  functions (wasm, scheduler), only_used_in_recursion on context
  params (cranelift, wasm, typeck), result_large_err on unification
- Mark future-use code with #[allow(dead_code)] (wasm BumpAllocator,
  interop deserialization structs)
- Auto-fixed: clone_on_copy, map_or simplification, redundant closure,
  useless format!, into_iter, is_multiple_of, elided lifetimes,
  Range::contains, collapsible if, derivable impl, let binding return

+++ compiler/eclexia-codegen/src/vm.rs
@@ -1154,12 +1154,12 @@ impl VirtualMachine {
                         Ok(ExecutionResult::Continue)
                     }
                     Value::Range(start, end) if field.as_str() == "len" => {
-                        let len = end.saturating_sub(start).max(0) as i64;
+                        let len = end.saturating_sub(start).max(0);
                         self.push(Value::Int(len))?;
                         Ok(ExecutionResult::Continue)
                     }
                     Value::RangeInclusive(start, end) if field.as_str() == "len" => {
-                        let len = end.saturating_sub(start).max(0) as i64 + 1;
+                        let len = end.saturating_sub(start).max(0) + 1;
                         self.push(Value::Int(len))?;
                         Ok(ExecutionResult::Continue)
                     }
